--- app/src/main/java/com/example/alldayfit/exercisestatus/BodyStatusEdit.kt
@@ -1,6 +1,0 @@
-package com.example.alldayfit.exercisestatus
-
-data class BodyStatusEdit (
-    var weight : String,
-    var height : String
-)
