--- app/src/main/java/com/example/alldayfit/exercisestatus/ExerciseStatusDailyEditDialog.kt
@@ -1,45 +1,0 @@
-package com.example.alldayfit.exercisestatus
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
-import com.example.alldayfit.databinding.ExerciseStatusDailyEditDialogBinding
-
-class ExerciseStatusDailyEditDialog : DialogFragment() {
-    private var _binding: ExerciseStatusDailyEditDialogBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = ExerciseStatusDailyEditDialogBinding.inflate(inflater, container, false)
-        initView()
-        return binding.root
-    }
-
-    /* dialog design, data 초기 설정 */
-    private fun initView() = with(binding) {
-        /*  아이템 text 초기 설정*/
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        setupView()
-    }
-
-    private fun setupView() = with(binding) {
-        closeBtn.setOnClickListener {
-            dismiss()
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-}
