+++ app/src/main/java/com/example/alldayfit/exercisestatus/model/BodyStatusEdit.kt
@@ -1,0 +1,6 @@
+package com.example.alldayfit.exercisestatus.model
+
+data class BodyStatusEdit (
+    var weight : Int,
+    var height : Int
+)
